backend/server/requirements.txt
fastapi
environs
asyncpg
uvicorn
pydantic
PyJWT
miniopy-async
librosa
soundfile
streaming-form-data
ollama
python-multipart